+++ addon/utils/select-box/focus.js
@@ -5,25 +5,23 @@ export function focusIn(selectBox, e) {
     selectBox.isFocused = true;
   } catch (error) {
     // https://github.com/emberjs/ember.js/issues/18043
-  } finally {
-    focusedIn(selectBox, e);
   }
 }
 
 export function focusOut(selectBox, e) {
   try {
     selectBox.isFocused = false;
+
+    focusedOut(selectBox, e);
   } catch (error) {
     // https://github.com/emberjs/ember.js/issues/18043
-  } finally {
-    focusedOut(selectBox, e);
   }
 }
 
-function focusedIn(selectBox, e) {
-  invokeAction(selectBox, 'onFocusIn', e, selectBox.api);
-}
-
 function focusedOut(selectBox, e) {
-  invokeAction(selectBox, 'onFocusOut', e, selectBox.api);
+  const focusLeave = !selectBox.element.contains(e.relatedTarget);
+
+  if (focusLeave) {
+    invokeAction(selectBox, 'onFocusLeave', e, selectBox.api);
+  }
 }