+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 15.0.0
+
+- Remove `onFocusIn` and `onFocusOut` events, these don't provide anything different that what can be acheived by using `{{on "focusin"}}` and `{{on "focusout"}}`.
+- Add `onFocusLeave` which fires when focus leaves the select box. This is useful for closing the select box when tabbing away.
+- Remove `onClickOutside` in favour of `onFocusLeave`. (This removed the need for event listeners on the document).
+
 ## 14.0.4
 
 - Upgrade dependencies